requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coverage==7.2.1
cron-descriptor==1.3.0
Django==4.1.7
django-celery-beat==2.5.0
django-ckeditor==6.5.1
django-inline-svg==0.1.1
django-js-asset==2.0.0
django-sites==0.11
django-timezone-field==5.0
django-tinymce==3.5.0
dnspython==2.3.0
eventlet==0.33.3
greenlet==2.0.2
gunicorn==20.1.0
kombu==5.2.4
mysqlclient==2.1.1
prompt-toolkit==3.0.38
psycopg2==2.9.5
PyMySQL==1.0.2
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.4
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
vine==5.0.0
wcwidth==0.2.6
whitenoise==6.4.0
